+++ python/ray/data/_internal/execution/operators/input_data_buffer.py
@@ -18,7 +18,7 @@ class InputDataBuffer(PhysicalOperator):
     def __init__(
         self,
         input_data: Optional[List[RefBundle]] = None,
-        input_data_factory: Callable[[int], List[RefBundle]] = None,
+        input_data_factory: Optional[Callable[[int], List[RefBundle]]] = None,
         num_output_blocks: Optional[int] = None,
     ):
         """Create an InputDataBuffer.
