+++ python/ray/data/_internal/logical/operators/input_data_operator.py
@@ -14,7 +14,7 @@ class InputData(LogicalOperator):
     def __init__(
         self,
         input_data: Optional[List[RefBundle]] = None,
-        input_data_factory: Optional[Callable[[], List[RefBundle]]] = None,
+        input_data_factory: Optional[Callable[[int], List[RefBundle]]] = None,
     ):
         assert (input_data is None) != (
             input_data_factory is None
